// src/census/enum/census-variable.enum.ts
export enum CensusVariable {
  TOTAL_POP = 'S0101_C01_001E',
  AGE_UNDER_5 = 'S0101_C01_002E',
  AGE_5T9 = 'S0101_C01_003E',
  AGE_10T14 = 'S0101_C01_004E',
  AGE_15T19 = 'S0101_C01_005E',
  TOTAL_POVERTY_STATUS_UNDER_6 = 'B17020_003E',
  TOTAL_POVERTY_STATUS_6_TO_11 = 'B17020_004E',
  TOTAL_POVERTY_STATUS_12_TO_17 = 'B17020_005E',
}